backend/requirements.txt
Django>=5.2,<6.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
django-ratelimit>=4.1
cryptography>=42.0
psycopg2-binary>=2.9
Pillow>=10.0
python-dotenv>=1.0
gunicorn>=21.0
